+++ app/astro/.eslintrc.cjs
@@ -1,15 +1,40 @@
+const path = require("path");
+
+/** @type {import("eslint").Linter.Config} */
 module.exports = {
   root: true,
-  extends: "custom",
+  extends: "@repo/eslint-config/index.cjs",
+  parser: "@typescript-eslint/parser",
   parserOptions: {
     tsconfigRootDir: __dirname,
     project: "./tsconfig.json",
   },
+  settings: {
+    "import/resolver": {
+      typescript: {
+        project: path.resolve(__dirname, "./tsconfig.json"),
+      },
+    },
+  },
   overrides: [
     {
-      files: ["astro.config.ts", "panda.config.ts"],
+      files: ["*.astro"],
+      parser: "astro-eslint-parser",
+      parserOptions: {
+        parser: "@typescript-eslint/parser",
+        extraFileExtensions: [".astro"],
+      },
+    },
+    {
+      files: ["astro.config.ts", "panda.config.ts", "postcss.config.cjs"],
       rules: {
         "import/no-extraneous-dependencies": "off",
+      },
+    },
+    {
+      files: ["postcss.config.cjs"],
+      rules: {
+        "global-require": "off",
       },
     },
     {