+++ app/astro/astro.config.ts
@@ -1,14 +1,14 @@
 import { defineConfig } from "astro/config";
-import pandacss from "@pandacss/astro";
+import icon from "astro-icon";
 import mdx from "@astrojs/mdx";
 import partytown from "@astrojs/partytown";
 import rehypeClassNames from "rehype-class-names";
-import { rehypeClassNamesOptions } from "@cieloazul310/astro-sarkara/classes";
+import mdxClasses from "./src/mdx-classes";
 
 // https://astro.build/config
 export default defineConfig({
   integrations: [
-    pandacss(),
+    icon(),
     mdx(),
     partytown({
       // Adds dataLayer.push as a forwarding-event.
@@ -18,6 +18,6 @@ export default defineConfig({
     }),
   ],
   markdown: {
-    rehypePlugins: [[rehypeClassNames, rehypeClassNamesOptions]],
+    rehypePlugins: [[rehypeClassNames, mdxClasses]],
   },
 });