+++ app/astro/src/mdx-classes.ts
@@ -1,0 +1,3 @@
+import { articleClasses } from "../styled-system/recipes";
+
+export default articleClasses();