+++ packages/components/.eslintrc.cjs
@@ -1,17 +1,36 @@
+const path = require("path");
+
+/** @type {import("eslint").Linter.Config} */
 module.exports = {
   root: true,
-  extends: "custom",
+  extends: "@repo/eslint-config/index.cjs",
+  parser: "@typescript-eslint/parser",
   parserOptions: {
     tsconfigRootDir: __dirname,
     project: "./tsconfig.json",
   },
-  ignorePatterns: ["dist"],
+  settings: {
+    "import/resolver": {
+      typescript: {
+        project: path.resolve(__dirname, "./tsconfig.json"),
+      },
+    },
+  },
   overrides: [
     {
-      files: ["src/*.ts"],
+      files: ["src/**/*.ts"],
       rules: {
         "import/no-extraneous-dependencies": "off",
       },
     },
+    {
+      files: ["*.astro"],
+      parser: "astro-eslint-parser",
+      parserOptions: {
+        parser: "@typescript-eslint/parser",
+        extraFileExtensions: [".astro"],
+      },
+    },
   ],
+  ignorePatterns: ["dist"],
 };