+++ packages/components/tsup.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  entry: ["src/preset/index.ts"],
+  external: ["@pandacss/dev"],
+  format: ["cjs", "esm"],
+  dts: true,
+  outDir: "dist",
+});