+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export * from "@cieloazul310/astro-sarkara-article-classes";
+// export * from "@cieloazul310/astro-sarkara-article-classes";
 export * from "@cieloazul310/astro-sarkara-components";
 export * from "@cieloazul310/astro-sarkara-layout";
 export * from "@cieloazul310/astro-sarkara-utils";