+++ packages/layout/.eslintrc.cjs
@@ -1,8 +1,29 @@
+const path = require("path");
+
+/** @type {import("eslint").Linter.Config} */
 module.exports = {
   root: true,
-  extends: "custom",
+  extends: "@repo/eslint-config/index.cjs",
+  parser: "@typescript-eslint/parser",
   parserOptions: {
     tsconfigRootDir: __dirname,
     project: "./tsconfig.json",
   },
+  settings: {
+    "import/resolver": {
+      typescript: {
+        project: path.resolve(__dirname, "./tsconfig.json"),
+      },
+    },
+  },
+  overrides: [
+    {
+      files: ["*.astro"],
+      parser: "astro-eslint-parser",
+      parserOptions: {
+        parser: "@typescript-eslint/parser",
+        extraFileExtensions: [".astro"],
+      },
+    },
+  ],
 };