+++ packages/layout/src/index.ts
@@ -1,4 +1,3 @@
-// @ts-nocheck
 export {
   default as SarkaraLayout,
   type Props as SarkaraLayoutProps,