+++ packages/layout/src/scripts/drawer.ts
@@ -1,8 +1,13 @@
-document.querySelectorAll(".drawer-handler")?.forEach((element) => {
-  element.addEventListener("click", () => {
-    const open = Boolean(
-      JSON.parse(document.body.getAttribute("data-drawer-open") ?? "false"),
-    );
-    document.body.setAttribute("data-drawer-open", JSON.stringify(!open));
+function setDrawerHandler() {
+  document.querySelectorAll(".drawer-handler")?.forEach((element) => {
+    element.addEventListener("click", () => {
+      const open = Boolean(
+        JSON.parse(document.body.getAttribute("data-drawer-open") ?? "false"),
+      );
+      document.body.setAttribute("data-drawer-open", JSON.stringify(!open));
+    });
   });
-});
+}
+
+setDrawerHandler();
+document.addEventListener("astro:after-swap", setDrawerHandler);