+++ packages/layout/src/scripts/toggleColorMode.ts
@@ -1,13 +1,13 @@
-document.querySelectorAll(".toggle-color-mode")?.forEach((element) => {
-  element.addEventListener("click", () => {
-    const colorMode = document.documentElement?.getAttribute("data-color-mode");
-    document.documentElement?.setAttribute(
-      "data-color-mode",
-      colorMode === "dark" ? "light" : "dark",
-    );
-    window.localStorage.setItem(
-      "astro-color-mode",
-      colorMode === "dark" ? "light" : "dark",
-    );
+function setDarkModeHandler() {
+  document.querySelectorAll(".color-mode-handler")?.forEach((element) => {
+    element.addEventListener("click", () => {
+      document.documentElement.classList.toggle("dark");
+
+      const isDark = document.documentElement.classList.contains("dark");
+      localStorage.setItem("theme", isDark ? "dark" : "light");
+    });
   });
-});
+}
+
+setDarkModeHandler();
+document.addEventListener("astro:after-swap", setDarkModeHandler);