+++ packages/preset-base/.eslintrc.cjs
@@ -1,7 +1,8 @@
+/** @type {import("eslint").Linter.Config} */
 module.exports = {
   root: true,
-  extends: "custom",
-  ignorePatterns: ["dist"],
+  extends: "@repo/eslint-config/index.cjs",
+  parser: "@typescript-eslint/parser",
   parserOptions: {
     tsconfigRootDir: __dirname,
     project: "./tsconfig.json",
@@ -10,4 +11,5 @@ module.exports = {
     "import/no-extraneous-dependencies": "off",
     "import/prefer-default-export": "off",
   },
+  ignorePatterns: ["dist"],
 };