+++ packages/preset-base/src/globalCss.ts
@@ -1,8 +1,8 @@
 import { defineGlobalStyles } from "@pandacss/dev";
 
 const body = {
-  backgroundColor: { base: "bg", _dark: "bg.dark" },
-  color: { base: "neutral.950", _dark: "neutral.50" },
+  backgroundColor: "bg",
+  color: "fg",
   "&[data-drawer-open=true]": {
     overflowY: "hidden",
   },